plusultra-DISMISC/src/requirements.txt
colorama
discord.py
aiohttp